test(favorit): modify modules favorit add for load new favorit structure
