+++ itachallenge-user/src/test/java/com/itachallenge/user/controller/UserControllerTest.java
@@ -8,6 +8,7 @@ import com.itachallenge.user.exception.NotFoundException;
 import com.itachallenge.user.exception.UserGlobalExceptionHandler;
 import com.itachallenge.user.service.IUserSolutionService;
 import com.itachallenge.user.service.UserService;
+import com.itachallenge.userinteraction.service.favorite.FavoriteService;
 import org.junit.jupiter.api.*;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -25,6 +26,9 @@ class UserControllerTest {
 
     @Mock
     private UserService userService;
+
+    @Mock
+    private FavoriteService favoriteService;
 
     @Mock
     private IUserSolutionService userSolutionService;
@@ -110,7 +114,7 @@ class UserControllerTest {
     void addToFavorites_WhenAdded_Returns201() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.just(true));
 
         webTestClient.post()
@@ -119,7 +123,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.CREATED)
                 .expectBody(Boolean.class).isEqualTo(true);
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -142,7 +146,7 @@ class UserControllerTest {
     void addToFavorites_WhenAlreadyInFavorites_Returns200() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.just(false));
 
         webTestClient.post()
@@ -151,7 +155,7 @@ class UserControllerTest {
                 .expectStatus().isOk()
                 .expectBody(Boolean.class).isEqualTo(false);
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -174,7 +178,7 @@ class UserControllerTest {
     void addToFavorites_WhenUserNotExists_Returns404() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new NotFoundException("User not found")));
 
         webTestClient.post()
@@ -183,7 +187,7 @@ class UserControllerTest {
                 .expectStatus().isNotFound()
                 .expectBody(String.class).isEqualTo("User not found");
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -206,7 +210,7 @@ class UserControllerTest {
     void addToFavorites_WhenBadFormattedId_Returns400() {
         String userId = "invalidUuid";
         String challengeId = "invalidUUid";
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new BadUUIDException("Error message")));
 
         webTestClient.post()
@@ -215,7 +219,7 @@ class UserControllerTest {
                 .expectStatus().isBadRequest()
                 .expectBody(String.class).isEqualTo("The provided IDs are not valid.");
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -238,7 +242,7 @@ class UserControllerTest {
     void addToFavorites_WhenUnexpectedError_Returns500() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new Exception()));
 
         webTestClient.post()
@@ -247,7 +251,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
                 .expectBody(String.class).isEqualTo("Unexpected error happened.");
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
